moved auth service to services folder. deleted sessions file
