+++ src/app/need-auth.guard.ts
@@ -1,6 +1,6 @@
 import {CanActivate, Router} from '@angular/router';
 import {Injectable} from '@angular/core';
-import {AuthService} from './auth/auth.service';
+import {AuthService} from './shared/services/auth.service';
 import {ActivatedRouteSnapshot, RouterStateSnapshot} from '@angular/router/src/router_state';
 
 @Injectable()
